+++ ionoscloud/resource_ipfailover.go
@@ -3,8 +3,14 @@ package ionoscloud
 import (
 	"context"
 	"fmt"
+	"github.com/ionos-cloud/terraform-provider-ionoscloud/v6/services/cloudapi/lanSvc"
 	"log"
 	"strings"
+
+	"github.com/ionos-cloud/terraform-provider-ionoscloud/v6/internal/uuidgen"
+	"github.com/ionos-cloud/terraform-provider-ionoscloud/v6/slice"
+	"github.com/ionos-cloud/terraform-provider-ionoscloud/v6/utils"
+	"github.com/ionos-cloud/terraform-provider-ionoscloud/v6/utils/constant"
 
 	"github.com/ionos-cloud/terraform-provider-ionoscloud/v6/services"
 	"github.com/ionos-cloud/terraform-provider-ionoscloud/v6/services/cloudapi"
@@ -28,16 +34,20 @@ func resourceLanIPFailover() *schema.Resource {
 			"ip": {
 				Type:             schema.TypeString,
 				Required:         true,
+				ForceNew:         true,
+				Description:      "Failover IP",
 				ValidateDiagFunc: validation.ToDiagFunc(validation.All(validation.IsIPAddress)),
 			},
 			"nicuuid": {
 				Type:             schema.TypeString,
 				Required:         true,
 				ValidateDiagFunc: validation.ToDiagFunc(validation.IsUUID),
+				Description:      "The UUID of the master NIC",
 			},
 			"lan_id": {
 				Type:             schema.TypeString,
 				Required:         true,
+				ForceNew:         true,
 				ValidateDiagFunc: validation.ToDiagFunc(validation.StringIsNotWhiteSpace),
 			},
 			"datacenter_id": {
@@ -53,28 +63,34 @@ func resourceLanIPFailover() *schema.Resource {
 
 func resourceLanIPFailoverCreate(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
 	client := meta.(services.SdkBundle).CloudApiClient
-	dcid := d.Get("datacenter_id").(string)
-	lanid := d.Get("lan_id").(string)
-	if lanid == "" {
-		diags := diag.FromErr(fmt.Errorf("'lan_id' is missing, please provide a valid lan ID "))
-		return diags
-	}
+	dcId := d.Get("datacenter_id").(string)
+	lanId := d.Get("lan_id").(string)
 	ip := d.Get("ip").(string)
 	nicUuid := d.Get("nicuuid").(string)
 
-	properties := &ionoscloud.LanProperties{}
-
-	properties.IpFailover = &[]ionoscloud.IPFailover{
-		{
-			Ip:      &ip,
-			NicUuid: &nicUuid,
-		}}
-
-	lan, apiResponse, err := client.LANsApi.DatacentersLansPatch(ctx, dcid, lanid).Lan(*properties).Execute()
-	logApiRequestTime(apiResponse)
+	// First, retrieve the existing IP Failover groups
+	lan, apiResponse, err := lanSvc.FindLanById(*client, ctx, dcId, lanId)
 	if err != nil {
-		diags := diag.FromErr(fmt.Errorf("an error occured while patching a lans failover group  %s %w", lanid, err))
-		return diags
+		if apiResponse.HttpNotFound() {
+			return diag.FromErr(fmt.Errorf("unable to find the LAN with ID: %s, datacenter ID: %s", lanId, dcId))
+		}
+		return diag.FromErr(err)
+	}
+	if lan.Properties == nil || lan.Properties.IpFailover == nil {
+		return diag.FromErr(fmt.Errorf("expected a LAN response containing IP failover groups but received 'nil' instead"))
+	}
+
+	// Add the new IP failover group to the list
+	*lan.Properties.IpFailover = append(*lan.Properties.IpFailover, ionoscloud.IPFailover{
+		Ip:      &ip,
+		NicUuid: &nicUuid,
+	})
+
+	// Modify the LAN using the new list
+	lan, apiResponse, err = client.LANsApi.DatacentersLansPatch(ctx, dcId, lanId).Lan(*lan.Properties).Execute()
+	apiResponse.LogInfo()
+	if err != nil {
+		return diag.FromErr(fmt.Errorf("an error occured while patching a lans IP failover group, LAN ID: %s, error: %w", lanId, err))
 	}
 
 	// Wait, catching any errors
@@ -84,56 +100,51 @@ func resourceLanIPFailoverCreate(ctx context.Context, d *schema.ResourceData, me
 		return diags
 	}
 
-	d.SetId(*lan.Id)
-
-	return resourceLanIPFailoverRead(ctx, d, meta)
+	// Use the IP in order to generate the resource ID
+	d.SetId(uuidgen.GenerateUuidFromName(ip))
+	return nil
 }
 
 func resourceLanIPFailoverRead(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
 	client := meta.(services.SdkBundle).CloudApiClient
+	dcId := d.Get("datacenter_id").(string)
+	lanId := d.Get("lan_id").(string)
+	ip := d.Get("ip").(string)
 
-	lan, apiResponse, err := client.LANsApi.DatacentersLansFindById(ctx, d.Get("datacenter_id").(string), d.Id()).Execute()
-	logApiRequestTime(apiResponse)
-
+	// Retrieve the existing IP Failover groups
+	lan, apiResponse, err := lanSvc.FindLanById(*client, ctx, dcId, lanId)
 	if err != nil {
-		if httpNotFound(apiResponse) {
+		if apiResponse.HttpNotFound() {
 			d.SetId("")
-			return nil
+			return diag.FromErr(fmt.Errorf("unable to find the LAN with ID: %s, datacenter ID: %s", lanId, dcId))
 		}
-		diags := diag.FromErr(fmt.Errorf("an error occured while fetching a lan ID %s %w", d.Id(), err))
-		return diags
+		return diag.FromErr(err)
+	}
+	if lan.Properties == nil || lan.Properties.IpFailover == nil {
+		return diag.FromErr(fmt.Errorf("expected a LAN response containing IP failover groups but received 'nil' instead"))
 	}
 
-	failoverSlice := lan.Properties.IpFailover
-	if lan.Properties != nil && failoverSlice != nil && len(*failoverSlice) > 0 {
-		firstFailover := (*failoverSlice)[0]
-		if firstFailover.Ip != nil {
-			err := d.Set("ip", firstFailover.Ip)
-			if err != nil {
-				diags := diag.FromErr(fmt.Errorf("error while setting ip property for IpFailover %s: %w", d.Id(), err))
-				return diags
+	// Iterate through IP Failover groups and select the proper one using the IP (the IP acts like
+	// an identifier) and then set the data accordingly.
+	ipFailoverGroups := lan.Properties.IpFailover
+	ipFailoverGroupFound := false
+	if lan.Properties != nil && ipFailoverGroups != nil && len(*ipFailoverGroups) > 0 {
+		for _, ipFailoverGroup := range *ipFailoverGroups {
+			if *ipFailoverGroup.Ip == ip {
+				if err := d.Set("nicuuid", *ipFailoverGroup.NicUuid); err != nil {
+					return diag.FromErr(utils.GenerateSetError(constant.ResourceIpFailover, "nicuuid", err))
+				}
+				ipFailoverGroupFound = true
+				break
 			}
 		}
-		if firstFailover.NicUuid != nil {
-			err := d.Set("nicuuid", firstFailover.NicUuid)
-			if err != nil {
-				diags := diag.FromErr(fmt.Errorf("error while setting nicuuid property for IpFailover %s: %w", d.Id(), err))
-				return diags
-			}
-		}
 	}
 
-	if lan.Id != nil {
-		err := d.Set("lan_id", *lan.Id)
-		if err != nil {
-			diags := diag.FromErr(fmt.Errorf("error while setting lan_id property for IpFailover %s: %w", d.Id(), err))
-			return diags
-		}
-	}
-
-	if err := d.Set("datacenter_id", d.Get("datacenter_id").(string)); err != nil {
-		diags := diag.FromErr(err)
-		return diags
+	// If the IP failover group was not found, set the ID to the empty string in order to remove the
+	// resource from the state.
+	if !ipFailoverGroupFound {
+		log.Printf("[INFO] IP Failover Group with IP: %s was not found in LAN with ID: %s, datacenter ID: %s", ip, lanId, dcId)
+		d.SetId("")
 	}
 
 	return nil
@@ -141,61 +152,85 @@ func resourceLanIPFailoverRead(ctx context.Context, d *schema.ResourceData, meta
 
 func resourceLanIPFailoverUpdate(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
 	client := meta.(services.SdkBundle).CloudApiClient
-
-	properties := &ionoscloud.LanProperties{}
-	dcid := d.Get("datacenter_id").(string)
-	lanid := d.Get("lan_id").(string)
+	dcId := d.Get("datacenter_id").(string)
+	lanId := d.Get("lan_id").(string)
 	ip := d.Get("ip").(string)
-	nicUuid := d.Get("nicuuid").(string)
 
-	properties.IpFailover = &[]ionoscloud.IPFailover{
-		{
+	if d.HasChange("nicuuid") {
+		oldValue, newValue := d.GetChange("nicuuid")
+		newNicUuid := newValue.(string)
+		oldNicUuid := oldValue.(string)
+
+		// First, retrieve the existing IP Failover groups
+		lan, apiResponse, err := lanSvc.FindLanById(*client, ctx, dcId, lanId)
+		if err != nil {
+			if apiResponse.HttpNotFound() {
+				return diag.FromErr(fmt.Errorf("unable to find the LAN with ID: %s, datacenter ID: %s", lanId, dcId))
+			}
+			return diag.FromErr(err)
+		}
+		if lan.Properties == nil || lan.Properties.IpFailover == nil {
+			return diag.FromErr(fmt.Errorf("expected a LAN response containing IP failover groups but received 'nil' instead"))
+		}
+
+		// Add the new IP failover group to the list
+		*lan.Properties.IpFailover = append(*lan.Properties.IpFailover, ionoscloud.IPFailover{
 			Ip:      &ip,
-			NicUuid: &nicUuid,
-		}}
+			NicUuid: &newNicUuid,
+		})
 
-	_, apiResponse, err := client.LANsApi.DatacentersLansPatch(ctx, dcid, lanid).Lan(*properties).Execute()
-	logApiRequestTime(apiResponse)
-	if err != nil {
-		diags := diag.FromErr(fmt.Errorf("an error occured while patching a lan ID %s %w", d.Id(), err))
-		return diags
+		// Remove the old IP failover group from the list
+		*lan.Properties.IpFailover = slice.DeleteFrom(*lan.Properties.IpFailover, ionoscloud.IPFailover{
+			Ip:      &ip,
+			NicUuid: &oldNicUuid,
+		})
+
+		_, apiResponse, err = client.LANsApi.DatacentersLansPatch(ctx, dcId, lanId).Lan(*lan.Properties).Execute()
+		apiResponse.LogInfo()
+		if err != nil {
+			diags := diag.FromErr(fmt.Errorf("an error occured while patching the lan with ID: %s, error: %w", lanId, err))
+			return diags
+		}
+
+		// Wait, catching any errors
+		_, errState := cloudapi.GetStateChangeConf(meta, d, apiResponse.Header.Get("Location"), schema.TimeoutUpdate).WaitForStateContext(ctx)
+		if errState != nil {
+			diags := diag.FromErr(errState)
+			return diags
+		}
 	}
-
-	// Wait, catching any errors
-	_, errState := cloudapi.GetStateChangeConf(meta, d, apiResponse.Header.Get("Location"), schema.TimeoutUpdate).WaitForStateContext(ctx)
-	if errState != nil {
-		diags := diag.FromErr(errState)
-		return diags
-	}
-
-	return resourceLanIPFailoverRead(ctx, d, meta)
+	return nil
 }
 
 func resourceLanIPFailoverDelete(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
 	client := meta.(services.SdkBundle).CloudApiClient
+	dcId := d.Get("datacenter_id").(string)
+	lanId := d.Get("lan_id").(string)
+	ip := d.Get("ip").(string)
+	nicUuid := d.Get("nicuuid").(string)
 
-	dcid := d.Get("datacenter_id").(string)
-	lanid := d.Get("lan_id").(string)
-
-	//remove the failover group
-	properties := &ionoscloud.LanProperties{
-		IpFailover: &[]ionoscloud.IPFailover{},
+	// First, retrieve the existing IP Failover groups
+	lan, apiResponse, err := lanSvc.FindLanById(*client, ctx, dcId, lanId)
+	if err != nil {
+		if apiResponse.HttpNotFound() {
+			return diag.FromErr(fmt.Errorf("unable to find the LAN with ID: %s, datacenter ID: %s", lanId, dcId))
+		}
+		return diag.FromErr(err)
+	}
+	if lan.Properties == nil || lan.Properties.IpFailover == nil {
+		return diag.FromErr(fmt.Errorf("expected a LAN response containing IP failover groups but received 'nil' instead"))
 	}
 
-	_, apiResponse, err := client.LANsApi.DatacentersLansPatch(ctx, dcid, lanid).Lan(*properties).Execute()
-	logApiRequestTime(apiResponse)
-	if err != nil {
-		/*
-						//try again in 90 seconds
-						time.Sleep(90 * time.Second)
-						_, apiResponse, err = client.LANsApi.DatacentersLansPatch(ctx, dcid, lanid).Lan(*properties).Execute()
-			logApiRequestTime(apiResponse)
+	// Remove the failover group from the list
+	*lan.Properties.IpFailover = slice.DeleteFrom(*lan.Properties.IpFailover, ionoscloud.IPFailover{
+		Ip:      &ip,
+		NicUuid: &nicUuid,
+	})
 
-						if err != nil && (apiResponse == nil || apiResponse.StatusCode != 404) {
-							return fmt.Errorf("an error occured while removing a lans ipfailover groups dcId %s ID %s %s", d.Get("datacenter_id").(string), d.Id(), err)
-						}
-		*/
-		diags := diag.FromErr(fmt.Errorf("an error occured while removing a lans ipfailover groups dcId %s ID %s %w", d.Get("datacenter_id").(string), d.Id(), err))
+	_, apiResponse, err = client.LANsApi.DatacentersLansPatch(ctx, dcId, lanId).Lan(*lan.Properties).Execute()
+	apiResponse.LogInfo()
+	if err != nil {
+		diags := diag.FromErr(fmt.Errorf("an error occured while removing an IP failover group with IP: %s for the LAN with ID: %s, datacenter ID: %s, error: %w", ip, lanId, dcId, err))
 		return diags
 	}
 
@@ -212,58 +247,53 @@ func resourceLanIPFailoverDelete(ctx context.Context, d *schema.ResourceData, me
 
 func resourceIpFailoverImporter(ctx context.Context, d *schema.ResourceData, meta interface{}) ([]*schema.ResourceData, error) {
 	parts := strings.Split(d.Id(), "/")
-	if len(parts) != 2 || parts[0] == "" || parts[1] == "" {
-		return nil, fmt.Errorf("invalid import id %q. Expecting {datacenter}/{lan}", d.Id())
+	if len(parts) != 3 || parts[0] == "" || parts[1] == "" || parts[2] == "" {
+		return nil, fmt.Errorf("invalid import ID: %s. Expecting {datacenter}/{lan}/{ip}", d.Id())
 	}
-
 	dcId := parts[0]
 	lanId := parts[1]
+	ip := parts[2]
 
 	client := meta.(services.SdkBundle).CloudApiClient
 
 	lan, apiResponse, err := client.LANsApi.DatacentersLansFindById(ctx, dcId, lanId).Execute()
-	logApiRequestTime(apiResponse)
+	apiResponse.LogInfo()
 
 	if err != nil {
-		if apiResponse != nil && apiResponse.Response != nil && apiResponse.StatusCode == 404 {
+		if apiResponse.HttpNotFound() {
 			d.SetId("")
-			return nil, fmt.Errorf("an error occured while trying to fetch the lan %q", lanId)
+			return nil, fmt.Errorf("unable to find the LAN with ID: %s, datacenter ID: %s", lanId, dcId)
 		}
-		return nil, fmt.Errorf("lan does not exist %q", lanId)
+		return nil, fmt.Errorf("error while fetching LAN with ID: %s, datacenter ID: %s, err: %w", lanId, dcId, err)
 	}
 
 	log.Printf("[INFO] lan found: %+v", lan)
 
-	d.SetId(*lan.Id)
+	ipFailoverGroups := lan.Properties.IpFailover
+	if lan.Properties != nil && ipFailoverGroups != nil && len(*ipFailoverGroups) > 0 {
+		for _, ipFailoverGroup := range *ipFailoverGroups {
+			// Search for the appropiate IP Failover Group using the provided IP
+			if *ipFailoverGroup.Ip == ip {
+				// Set all the information only if the IP Failover Group exists
+				// Use the IP in order to generate the resource ID
+				d.SetId(uuidgen.GenerateUuidFromName(ip))
 
-	if err := d.Set("datacenter_id", dcId); err != nil {
-		return nil, err
-	}
-
-	failoverSlice := lan.Properties.IpFailover
-	if lan.Properties != nil && failoverSlice != nil && len(*failoverSlice) > 0 {
-		firstFailover := (*failoverSlice)[0]
-		if firstFailover.Ip != nil {
-			err := d.Set("ip", firstFailover.Ip)
-			if err != nil {
-				return nil, fmt.Errorf("error while setting ip property for IpFailover %s: %w", d.Id(), err)
-
+				if err := d.Set("datacenter_id", dcId); err != nil {
+					return nil, utils.GenerateSetError(constant.ResourceIpFailover, "datacenter_id", err)
+				}
+				if err := d.Set("lan_id", lanId); err != nil {
+					return nil, utils.GenerateSetError(constant.ResourceIpFailover, "lan_id", err)
+				}
+				if err := d.Set("ip", ip); err != nil {
+					return nil, utils.GenerateSetError(constant.ResourceIpFailover, "ip", err)
+				}
+				if err := d.Set("nicuuid", *ipFailoverGroup.NicUuid); err != nil {
+					return nil, utils.GenerateSetError(constant.ResourceIpFailover, "nicuuid", err)
+				}
+				// After we find the IP Failover Group, we can stop searching since the IP is unique
+				return []*schema.ResourceData{d}, nil
 			}
 		}
-		if firstFailover.NicUuid != nil {
-			err := d.Set("nicuuid", firstFailover.NicUuid)
-			if err != nil {
-				return nil, fmt.Errorf("error while setting nicuuid property for IpFailover %s: %w", d.Id(), err)
-			}
-		}
 	}
-
-	if lan.Id != nil {
-		err := d.Set("lan_id", *lan.Id)
-		if err != nil {
-			return nil, err
-		}
-	}
-
-	return []*schema.ResourceData{d}, nil
+	return nil, fmt.Errorf("IP Failover Group with IP: %s does not exist in the LAN with ID: %s, datacenter ID: %s", ip, lanId, dcId)
 }